Basis for text fragments processing Strings are extarcted from SVG file
Methods calculating distance between a point and a line segment
(so far, there were methods for calculating distance between point
and an infinite line)  ... the new methods will be useful to find
the text fragment closest to the given point.

Bug-fixes for the intersection point detection. (In the new
vector implementation we were not taking into account special cases)