Much better rematching of labels with major ticks
If the ration between the number of matched labels and the number of major ticks
falls into a predefined interval, we match labels with only major ticks.

Labels which are too distant from the projection of average centre, on one of the axis, are rejected (In the case of having more labels than major ticks)
If there are more major ticks than labels, some are skipped and assignment is chosen in a way which minimises the sym of distances between ticks and labels.